Added product Repository,product mapper, product and base entity to coonect and call call to Database. Added Data base properties and configured hibernate and JPA and make network call, removed FakeStoreSingleProduct class as it was of no use

+++ src/main/java/com/example/Ecommerce/gateway/FakeStoreProductGateway.java
@@ -18,7 +18,7 @@ public class FakeStoreProductGateway implements IProductGateway{
         this.fakeStoreProductApi = fakeStoreProductApi;
     }
     @Override
-    public Map<Integer , ProductDTO> getAllProducts() throws IOException {
+    public Map<Long, ProductDTO> getAllProducts() throws IOException {
 
         FakeStoreProductResponseDTO response = this.fakeStoreProductApi.getAllFakeProducts().execute().body();
